+++ internal/constants/constants.go
@@ -58,7 +58,6 @@ const (
 	//
 	// Filenames.
 	//
-	StateFilename           = "constellation-state.json"
 	ClusterIDsFileName      = "constellation-id.json"
 	ConfigFilename          = "constellation-conf.yaml"
 	LicenseFilename         = "constellation.license"