+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 <!-- For now removed features. -->
 - `endpoint` flag of `constellation init`. IP is now always taken from the `constellation-id.json` file.
+- `constellation-state.json` file won't be created anymore. Resources are now managed through Terraform.
 
 ### Fixed